Implemented backend connection
Modified the App.jsx file, so it can request and update the number of counts through the backend server. The endpoint of the backend server must be given through as a environment variable.

TODO: Write a shell script which creates a S3 bucket for static web hosting, and hosts the production build of this React application.